Move schema to top level
This keeps all the ent-generated stuff hidden away where
you'll never need to look.
