+++ generate.go
@@ -1,5 +1,5 @@
 package puzzad
 
-//go:generate go run -mod=mod entgo.io/ent/cmd/ent generate ./ent/schema
+//go:generate go run -mod=mod entgo.io/ent/cmd/ent generate --target ./ent ./schema
 
 // segfaults, so... no. go:generate go run -mod=mod entgo.io/ent/cmd/ent describe ./ent/schema
