+++ schema/adventure.go
